Move base package installer classes definitions to their own module
