+++ napari_plugin_manager/_tests/test_qt_plugin_dialog.py
@@ -22,7 +22,7 @@ if qtpy.API_NAME == 'PySide2' and sys.version_info[:2] > (3, 10):
     )
 
 from napari_plugin_manager import qt_plugin_dialog
-from napari_plugin_manager.qt_package_installer import InstallerActions
+from napari_plugin_manager.base_qt_package_installer import InstallerActions
 
 N_MOCKED_PLUGINS = 2
